refact: UserResponseDto
네이밍 변경, 경로 변경
